backend/ai_agent/__init__.py
"""
LangGraph AI Agent for HCP CRM
Implements 5 tools: log, edit, query, summarize, extract_entities
"""
from typing import Any, Dict, List, Optional
from pydantic import BaseModel


class LogInteractionTool:
    """Tool to log HCP interactions"""
    
    @staticmethod
    def execute(hcp_name: str, interaction_details: str) -> Dict[str, Any]:
        return {
            'status': 'success',
            'message': f'Logged interaction with {hcp_name}',
            'interaction_id': 1,
            'summary': interaction_details[:100]
        }


class EditInteractionTool:
    """Tool to edit existing interactions"""
    
    @staticmethod
    def execute(interaction_id: int, updates: Dict[str, Any]) -> Dict[str, Any]:
        return {
            'status': 'success',
            'message': f'Updated interaction {interaction_id}',
            'updated_fields': list(updates.keys())
        }


class QueryInteractionsTool:
    """Tool to query interactions"""
    
    @staticmethod
    def execute(query: str, filters: Optional[Dict] = None) -> Dict[str, Any]:
        return {
            'status': 'success',
            'message': f'Found interactions matching: {query}',
            'count': 0,
            'interactions': []
        }


class SummarizeInteractionTool:
    """Tool to summarize interactions using AI"""
    
    @staticmethod
    def execute(interaction_id: int) -> Dict[str, Any]:
        return {
            'status': 'success',
            'message': f'Generated summary for interaction {interaction_id}',
            'summary': 'Interaction summary would be generated here'
        }


class ExtractEntitiesTool:
    """Tool to extract entities from interaction data"""
    
    @staticmethod
    def execute(interaction_text: str) -> Dict[str, Any]:
        return {
            'status': 'success',
            'message': 'Extracted entities from interaction',
            'entities': {
                'medications': [],
                'conditions': [],
                'procedures': [],
                'locations': [],
                'people': []
            }
        }


class HCPCRMAgent:
    """Main AI Agent for HCP CRM using LangGraph"""
    
    def __init__(self):
        self.tools = {
            'log_interaction': LogInteractionTool(),
            'edit_interaction': EditInteractionTool(),
            'query_interactions': QueryInteractionsTool(),
            'summarize_interaction': SummarizeInteractionTool(),
            'extract_entities': ExtractEntitiesTool(),
        }
    
    def process_message(self, message: str) -> Dict[str, Any]:
        """Process a message and route to appropriate tool"""
        message_lower = message.lower()
        
        if any(word in message_lower for word in ['log', 'create', 'record']):
            return self.tools['log_interaction'].execute(
                hcp_name='Unknown',
                interaction_details=message
            )
        elif any(word in message_lower for word in ['edit', 'update', 'change']):
            return self.tools['edit_interaction'].execute(
                interaction_id=1,
                updates={'notes': message}
            )
        elif any(word in message_lower for word in ['query', 'find', 'search', 'list']):
            return self.tools['query_interactions'].execute(query=message)
        elif any(word in message_lower for word in ['summarize', 'summary', 'recap']):
            return self.tools['summarize_interaction'].execute(interaction_id=1)
        elif any(word in message_lower for word in ['extract', 'entities', 'parse']):
            return self.tools['extract_entities'].execute(interaction_text=message)
        else:
            return {
                'status': 'processing',
                'message': f'Processing: {message}',
                'response': 'I can help you log, edit, query, summarize, or analyze HCP interactions.'
            }


# Initialize agent
agent = HCPCRMAgent()


def process_user_input(user_input: str) -> str:
    """Process user input and return response"""
    result = agent.process_message(user_input)
    return result.get('message', 'Processing your request...')
